fix: deleting files, correcting routes and controllers

+++ src/controllers/user.controller.js
@@ -1,60 +1,100 @@
 import { User } from "../models/user.models.js"
-import { UserService } from "../services/user/user.service.js"
 import { ERROR, SUCCESS } from "../shared/messages.js"
 
-const instanceServiceUser = new UserService();
-
 const createUser = async (req, res) => {
-    const { name, email, password } = req.body;
-    const newUser = await instanceServiceUser.createUserService(name, email, password);
-    return res
-        .status(201)
-        .json({
-            message: "Novo usuário criado com sucesso!",
-            newUser
+
+    const { name, email } = req.body
+    try {
+
+        const userAlreadyExist = await User.findOne({
+            where: {
+                name,
+                email
+            }
         })
+
+        if (userAlreadyExist) {
+            return res
+                .status(400)
+                .json({ error: `Usuário ${ERROR.ALREADY_EXIST}` })
+        }
+
+        const newUser = await User.create(req.body);
+        res.status(201).json(newUser);
+
+    } catch (error) {
+        return error
+    }
+
 }
 
 const getAllUser = async (req, res) => {
-    const users = instanceServiceUser.getAllUserService();
+    const users = await User.findAll();
     return res.json({ users });
 }
 
 const getUserByName = async (req, res) => {
+
     const { name } = req.body;
     const UserFindName = await User.findOne({
         where: {
             name
         }
     })
+
     return res.json({ UserFindName })
 }
 
 const updatePassword = async (req, res) => {
+
     const { id } = req.params;
-    const { newPassword } = req.body;
+    const { password: newPassword } = req.body;
 
-    const userAlreadyExist = await User.findOne({
-        where: {
-            id
+    try {
+        const user = await User.findOne({
+            where: {
+                id
+            }
+        })
+
+        if (!user) {
+
+            return res
+                .status(404)
+                .json({ message: `Usuario ${ERROR.NOT_FOUND}` })
         }
-    })
 
-    if (userAlreadyExist) {
-        return res.json({ message: `Usuário ${ERROR.NOT_FOUND}` })
+        await User.update({ password: newPassword }, {
+            where: {
+                id
+            }
+        })
+
+        const updatedUser = await User.findByPk(id)
+        return res.json({ message: `senha atualizada com sucesso!`, user: updatedUser })
+
+    } catch (error) {
+        return res
+            .status(500)
+            .json({ message: 'Erro no serivdor', error })
     }
-
-    await User.update({ password: newPassword }, {
-        where: {
-            id
-        }
-    })
-    const messageUpdate = await User.findByPk(id)
-    return res.json({ messageUpdate })
-}
+};
 
 const deleteUser = async (req, res) => {
-    const { id } = req.params
+
+    const { id } = req.params;
+    const user = await User.findOne({
+        where: {
+            id
+        }
+    })
+
+    if (!user) {
+        return res
+            .status(404)
+            .json({ message: `Usuario ${ERROR.NOT_FOUND}` })
+    }
+
     await User.destroy({
         where: {
             id
@@ -64,5 +104,6 @@ const deleteUser = async (req, res) => {
         message: `Usuário ${SUCCESS.DELETED}`
     })
 }
+
 
 export { createUser, getAllUser, getUserByName, updatePassword, deleteUser }
